Create CrazyTest Gateway
Realized that I needed to create unique devices for the second system because the "howmany" for-loop / ZMP process would choke on figuring out which ZMP applied to which set of devices.  I'm not 100% sure that's the case now that I rubber chicken it but nonetheless, In doing this, I am able to get rid of the tedium of entering S2 into every S2 device.  There is now a section for it.

I had a bunch of If-Else stuff going on because I was going to have it process differently if it was an S2 device but once there was a different section, then there was  a single If-Else after the initialization.  

I still haven't done the ZMP part that triggered this, and I'm sure there are better ways to do what I did so may be you can see something that can be optimized.  I took out the stuff that set S2 and all of that and just defined the variable you are using as S2 a couple of on line 86 and then got scared and left line 85 alone but I think it should be deleted.

I'll work on the ZMP stuff later.  But at least I know what it does now.
I just installed it and it came up ok.. didn't save.. got scared.  I'll do it later
I think it's OK but you should install at your own risk until I test.  :)